Web-Fundamentals/Python/news-python-system/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
email-validator
python-multipart
logging